+++ packages/cheeket/lib/provider/in-container-scope.ts
@@ -5,6 +5,7 @@ import Provider from "./provider";
 import { ContainerEventEmitter, DefaultState } from "../context";
 import { Middleware } from "../middleware";
 import bindInContext from "./bind-in-context";
+import emitCreateEvent from "./emit-create-event";
 
 function inContainerScope<T, State = DefaultState>(
   provider: Provider<T>,
@@ -38,9 +39,7 @@ function inContainerScope<T, State = DefaultState>(
           cache.set(context.container.id, value);
 
           context.container.addListener("close", handleOnClose);
-
-          context.container.emit("create", value);
-          await context.container.emitAsync("create:async", value);
+          await emitCreateEvent(context.container, value);
         }
 
         bindInContext(context, value, options);
